Add files via upload
I'm still trying to figure out what the hell to do with the Transformer.  I mean, I feel that the concept of word embeddings is only part of the whole story.  When we do brain scans of people who say or hear/understand a word, numerous regions light up, including smell, taste, feel, sight, memories, etc.  Is it possible to create a network a la Tononi's IIT model, with connections between equivalent embeddings?